Add orientation video to batch details
The code introduces an orientation video section in the batch detail information. A new dialog component is also included that appears on clicking the video. An IntersectionObserver is used to watch elements of the page and update the active tab based on user scrolling. This will make the interface more interactive and easily navigable. Lastly, some minor formatting and style adjustments have been incorporated for better readability and structure.