Adjust batch loading logic in BatchGridList.tsx
Updated the conditional check for fetched batches in BatchGridList component to either show or hide the "load more" button. Previously, the button would hide only when no more data was found, now, it will also be hidden when less than 6 batches are fetched. This change improves the user experience by appropriately managing the visibility of the "load more" button.