Change authentication redirect to open in same tab
The code modifications in the Navbar.tsx and LoginButton.tsx files reflect a change in the authentication process. Now, instead of opening in a new tab, the '/study/auth' will open in the same tab. This modification serves to streamline the login process and prevent the opening of unnecessary tabs.