Update logs and environment variables, modify SchoolCurriculumListing images
This commit includes alterations to logs in index.tsx and MainBanner, and the removal of a console.log in the deprecated index.tsx file. There are also updates to the environment variables NEXT_PUBLIC_HOMEPAGE_APP_ID and NEXT_PUBLIC_JEE_FAQ_ID in the .env file. SchoolCurriculumListing images and related variable names in schoolsAllclassDatajson.tsx have also been adjusted for better clarity.