Uncomment sections in index.tsx file
Sections related to HeroSection, HeroFeatureSection, etc were uncommented in the index.tsx file that was in the deprecated directory. Now, these components will be active and perform as intended. Also, this means the previous disabling of related parts for testing or debugging purposes is no longer active, returning the code to its functional state.