Commit message: Comment out Hero Section in index.tsx
The Hero Section has been commented out in the index.tsx file located in the deprecated directory. This action temporarily disables the effect of this section in the overall functionality of the application. This can be done to prevent its effects during debugging or testing other areas of the application.