Refactor Navbar and MainBanner, add new Logo and NextImage components
Replaced PwLogoInverted with a new Logo component in the Navbar and Image with NextImage in the MainBanner. Tempoarily commented out Login/Register buttons in Navbar and reinstated MainBanner display in index.tsx. New Logo and NextImage components were created to enhance code reusability and improve UI consistency across the platform.