Updated SEO and Layout components to enforce data requirement
Made changes in the SEO and Layout components to enforce the requirement of certain parameters. This step was taken to improve the predictability and reliability of the application. By ensuring these properties always have values, we eliminate potential errors and improve overall code readability. Removed unnecessary console logs in the Layout file.